Split up lib/formatting
