+++ client/lib/formatting/decode/browser.js
