+++ client/lib/formatting/decode/node.js
