+++ client/lib/formatting/decode/test/browser.js
