+++ client/lib/formatting/package.json
@@ -1,0 +1,3 @@
+{
+	"sideEffects": false
+}
